Completed re-usable function refactoring successfully :D. enhanced pattern designs DRY/KiSS. kept all features of error handling.
