+++ tests/2-profile.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect } from '@playwright/test';
-import { loginUser } from './utils/test-config';
+import { loginUser } from './utils/login';
 
 test.describe('Profile', () => {
   test('Update user profile information', async ({ page }) => {
