+++ tests/3-products-and-cart.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect } from '@playwright/test';
-import { loginUser } from './utils/test-config';
+import { loginUser } from './utils/login';
 
 test.describe('Products Shop & Cart Testing', () => {
